+++ app/src/main/java/com/saddict/rentalfinder/rentals/data/local/dao/ManageRentalDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.saddict.rentalfinder.rentals.model.local.RentalManageEntity
+import com.saddict.rentalfinder.rentals.model.local.rentals.RentalManageEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
