+++ app/src/main/java/com/saddict/rentalfinder/rentals/data/local/dao/RentalManageRemoteKeysDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.saddict.rentalfinder.rentals.model.local.RentalManageRemoteKeysEntity
+import com.saddict.rentalfinder.rentals.model.local.rentals.RentalManageRemoteKeysEntity
 
 @Dao
 interface RentalManageRemoteKeysDao {
