+++ app/src/main/java/com/saddict/rentalfinder/rentals/data/local/dao/RentalRemoteKeysDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.saddict.rentalfinder.rentals.model.local.RentalRemoteKeysEntity
+import com.saddict.rentalfinder.rentals.model.local.rentals.RentalRemoteKeysEntity
 
 @Dao
 interface RentalRemoteKeysDao {
