+++ app/src/main/java/com/saddict/rentalfinder/rentals/data/repository/Repository.kt
@@ -10,9 +10,9 @@ import com.saddict.rentalfinder.rentals.data.manager.ImageRemoteMediator
 import com.saddict.rentalfinder.rentals.data.manager.RentalManageRemoteMediator
 import com.saddict.rentalfinder.rentals.data.manager.RentalRemoteMediator
 import com.saddict.rentalfinder.rentals.data.remote.remository.RemoteDataSource
-import com.saddict.rentalfinder.rentals.model.local.ImageEntity
-import com.saddict.rentalfinder.rentals.model.local.RentalEntity
-import com.saddict.rentalfinder.rentals.model.local.RentalManageEntity
+import com.saddict.rentalfinder.rentals.model.local.images.ImageEntity
+import com.saddict.rentalfinder.rentals.model.local.rentals.RentalEntity
+import com.saddict.rentalfinder.rentals.model.local.rentals.RentalManageEntity
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
